# 8302152: Speed up tests with infinite loops, sleep less