--- a/test/hotspot/jtreg/compiler/loopopts/TestCMoveWithDeadPhi.java
+++ b/test/hotspot/jtreg/compiler/loopopts/TestCMoveWithDeadPhi.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2021, Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2021, 2023, Oracle and/or its affiliates. All rights reserved.
  * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
  *
  * This code is free software; you can redistribute it and/or modify it