--- a/test/hotspot/jtreg/compiler/loopopts/TestInfiniteLoopCCP.java
+++ b/test/hotspot/jtreg/compiler/loopopts/TestInfiniteLoopCCP.java
@@ -72,6 +72,6 @@ public class TestInfiniteLoopCCP {
         // Give thread some time to trigger compilation
         thread.setDaemon(true);
         thread.start();
-        Thread.sleep(Utils.adjustTimeout(4000));
+        Thread.sleep(Utils.adjustTimeout(500));
     }
 }